upgrade ldap WithLdapServer UTs
