+++ kyuubi-server/src/test/scala/org/apache/kyuubi/operation/thrift/http/KyuubiOperationThriftHttpKerberosAndPlainAuthSuite.scala
@@ -56,6 +56,10 @@ class KyuubiOperationThriftHttpKerberosAndPlainAuthSuite
         classOf[UserDefineAuthenticationProviderImpl].getCanonicalName)
       .set(KyuubiConf.SERVER_SPNEGO_KEYTAB, testKeytab)
       .set(KyuubiConf.SERVER_SPNEGO_PRINCIPAL, testSpnegoPrincipal)
+      .set(KyuubiConf.AUTHENTICATION_LDAP_GUIDKEY, ldapGuidKey)
+      .set(KyuubiConf.AUTHENTICATION_LDAP_PASSWORD, ldapBindnpw)
+      .set(KyuubiConf.AUTHENTICATION_LDAP_DOMAIN, ldapDomain)
+      .set(KyuubiConf.AUTHENTICATION_LDAP_ATTRIBUTES, ldapAttrs)
   }
 
   override protected def getJdbcUrl: String =
